requirements.txt
tefas-crawler==0.3.1
supabase-py==0.0.2
